EWPP-901: A bit of test cleanup.
