+++ tests/modules/oe_link_lists_test/src/Plugin/LinkDisplay/Foo.php
@@ -4,7 +4,6 @@ namespace Drupal\oe_link_lists_test\Plugin\LinkDisplay;
 
 use Drupal\Core\Link;
 use Drupal\oe_link_lists\LinkCollectionInterface;
-use Drupal\oe_link_lists\LinkDisplayPluginBase;
 
 /**
  * Plugin implementation of the link_display.
@@ -16,7 +15,7 @@ use Drupal\oe_link_lists\LinkDisplayPluginBase;
  *   bundles = { "dynamic", "manual" }
  * )
  */
-class Foo extends LinkDisplayPluginBase {
+class Foo extends GenericDisplayBase {
 
   /**
    * {@inheritdoc}
